# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Faturação",
        "description": "Documentação automática gerada pelo Swagger para a API de Faturação.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/backup/run": {
            "post": {
                "tags": [
                    "Backup"
                ],
                "summary": "Executa o backup do sistema",
                "description": "Executa o comando de backup do sistema utilizando o Artisan. Retorna uma mensagem de sucesso ou erro.",
                "operationId": "2f513fc93355ab420857cf9e56e3ecdf",
                "responses": {
                    "200": {
                        "description": "Backup criado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Backup criado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar backup.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar backup."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/backup/logs": {
            "get": {
                "tags": [
                    "Backup"
                ],
                "summary": "Obtém o arquivo de log do sistema",
                "description": "Retorna o arquivo de log principal do sistema (laravel.log) se existir, ou uma mensagem informando que não há logs.",
                "operationId": "a36d0c95695dd6df908db13c5afc4b84",
                "responses": {
                    "200": {
                        "description": "Arquivo de log retornado com sucesso."
                    },
                    "404": {
                        "description": "Nenhum log encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nenhum log encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos os clientes",
                "operationId": "f8eafbc2e9500413091c7f84a4470b84",
                "responses": {
                    "200": {
                        "description": "Lista de clientes retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Criar um novo cliente",
                "operationId": "2a8a28c7aec782e1af95193a65053e68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "nif": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "provincia": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "localizacao": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exibir um cliente específico",
                "operationId": "7be647ff4c2ab71c5a1f042133b13428",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente retornado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar um cliente existente",
                "operationId": "1cc735d14964cb4b25557d546e807a9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "nif": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "provincia": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "localizacao": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "422": {
                        "description": "Dados de entrada inválidos"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Remover um cliente",
                "operationId": "fae533cf0841c3ece1414287d89ef42b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
...
    },
...
}